+++ session/errors.go
@@ -79,6 +79,7 @@ const (
 	ER_TOO_MANY_KEY_PARTS
 	ER_TOO_LONG_IDENT
 	ER_UDPATE_TOO_MUCH_ROWS
+	ER_INSERT_TOO_MUCH_ROWS
 	ER_WRONG_NAME_FOR_INDEX
 	ER_TOO_MANY_KEYS
 	ER_NOT_SUPPORTED_KEY_TYPE
@@ -93,7 +94,6 @@ const (
 	ER_TABLE_MUST_INNODB
 	ER_TABLE_CHARSET_MUST_UTF8
 	ER_TABLE_CHARSET_MUST_NULL
-	ER_NAMES_MUST_UTF8
 	ER_TABLE_MUST_HAVE_COMMENT
 	ER_COLUMN_HAVE_NO_COMMENT
 	ER_TABLE_MUST_HAVE_PK
@@ -177,6 +177,7 @@ const (
 	ER_WRONG_TABLE_NAME
 	ER_CANT_SET_CHARSET
 	ER_CANT_SET_COLLATION
+	ER_MUST_AT_LEAST_ONE_COLUMN
 	ER_MUST_HAVE_COLUMNS
 	ER_PRIMARY_CANT_HAVE_NULL
 	ErrCantRemoveAllFields
@@ -185,6 +186,10 @@ const (
 	ErrNotFoundMasterStatus
 	ErrNonUniqTable
 	ErrWrongUsage
+	ErrDataTooLong
+	ErrCharsetNotSupport
+	ErrCollationNotSupport
+	ErrTableCollationNotSupport
 	ER_ERROR_LAST
 )
 
@@ -229,7 +234,8 @@ var ErrorsDefault = map[int]string{
 	ER_FOREIGN_KEY:                         "Foreign key is not allowed in table '%s'.",
 	ER_TOO_MANY_KEY_PARTS:                  "Too many key parts in Key '%s' in table '%s' specified, max %d parts allowed.",
 	ER_TOO_LONG_IDENT:                      "Identifier name '%s' is too long.",
-	ER_UDPATE_TOO_MUCH_ROWS:                "Update rows more then %d.",
+	ER_UDPATE_TOO_MUCH_ROWS:                "Update(%d) more then %d rows.",
+	ER_INSERT_TOO_MUCH_ROWS:                "Insert(%d) more then %d rows.",
 	ER_WRONG_NAME_FOR_INDEX:                "Incorrect index name '%s' in table '%s'.",
 	ER_TOO_MANY_KEYS:                       "Too many keys specified in table '%s', max %d keys allowed.",
 	ER_NOT_SUPPORTED_KEY_TYPE:              "Not supported key type: '%s'.",
@@ -244,7 +250,7 @@ var ErrorsDefault = map[int]string{
 	ER_TABLE_MUST_INNODB:                   "Set engine to innodb for table '%s'.",
 	ER_TABLE_CHARSET_MUST_UTF8:             "Set charset to one of '%s' for table '%s'.",
 	ER_TABLE_CHARSET_MUST_NULL:             "Not allowed set charset for table '%s'.",
-	ER_NAMES_MUST_UTF8:                     "Set charset to one of '%s'.",
+	ErrTableCollationNotSupport:            "Not allowed set collation for table '%s'.",
 	ER_TABLE_MUST_HAVE_COMMENT:             "Set comments for table '%s'.",
 	ER_COLUMN_HAVE_NO_COMMENT:              "Column '%s' in table '%s' have no comments.",
 	ER_TABLE_MUST_HAVE_PK:                  "Set a primary key for table '%s'.",
@@ -328,7 +334,8 @@ var ErrorsDefault = map[int]string{
 	ER_WRONG_TABLE_NAME:                    "Incorrect table name '%-.100s'",
 	ER_CANT_SET_CHARSET:                    "Cannot set charset '%s'",
 	ER_CANT_SET_COLLATION:                  "Cannot set collation '%s'",
-	ER_MUST_HAVE_COLUMNS:                   "A table must have at least 1 column",
+	ER_MUST_AT_LEAST_ONE_COLUMN:            "A table must have at least 1 column.",
+	ER_MUST_HAVE_COLUMNS:                   "Must have the specified column: '%s'.",
 	ER_PRIMARY_CANT_HAVE_NULL:              "All parts of a PRIMARY KEY must be NOT NULL; if you need NULL in a key, use UNIQUE instead",
 	ErrCantRemoveAllFields:                 "You can't delete all columns with ALTER TABLE; use DROP TABLE instead",
 	ErrNotFoundTableInfo:                   "Skip backup because there is no table structure information.",
@@ -336,6 +343,9 @@ var ErrorsDefault = map[int]string{
 	ErrNotFoundMasterStatus:                "Can't found master binlog position.",
 	ErrNonUniqTable:                        mysql.MySQLErrName[mysql.ErrNonuniqTable],
 	ErrWrongUsage:                          "Incorrect usage of %s and %s",
+	ErrDataTooLong:                         "Data too long for column '%s' at row %d",
+	ErrCharsetNotSupport:                   "Set charset to one of '%s'.",
+	ErrCollationNotSupport:                 "Set collation to one of '%s'",
 	ER_ERROR_LAST:                          "TheLastError,ByeBye",
 }
 
@@ -379,7 +389,8 @@ var ErrorsChinese = map[int]string{
 	ER_FOREIGN_KEY:                      "不允许使用外键(表 '%s').",
 	ER_TOO_MANY_KEY_PARTS:               "索引 '%s'指定了太多的字段(表 '%s'), 最多允许 %d 个字段.",
 	ER_TOO_LONG_IDENT:                   "名称 '%s' 过长.",
-	ER_UDPATE_TOO_MUCH_ROWS:             "更新行超过 %d .",
+	ER_UDPATE_TOO_MUCH_ROWS:             "一次更新(%d)超过 %d 行.",
+	ER_INSERT_TOO_MUCH_ROWS:             "一次新增(%d)超过 %d 行.",
 	ER_WRONG_NAME_FOR_INDEX:             "索引 '%s' 名称不正确(表 '%s').",
 	ER_TOO_MANY_KEYS:                    "表 '%s' 指定了太多索引, 最多允许 %d 个.",
 	ER_NOT_SUPPORTED_KEY_TYPE:           "不允许的键类型: '%s'.",
@@ -394,7 +405,7 @@ var ErrorsChinese = map[int]string{
 	ER_TABLE_MUST_INNODB:                   "仅支持innodb存储引擎(表'%s').",
 	ER_TABLE_CHARSET_MUST_UTF8:             "允许的字符集为: '%s'(表'%s').",
 	ER_TABLE_CHARSET_MUST_NULL:             "表 '%s' 禁止设置字符集!",
-	ER_NAMES_MUST_UTF8:                     "允许的字符集为: '%s'.",
+	ErrTableCollationNotSupport:            "表 '%s' 禁止设置排序规则!",
 	ER_TABLE_MUST_HAVE_COMMENT:             "表 '%s' 需要设置注释.",
 	ER_COLUMN_HAVE_NO_COMMENT:              "列 '%s' 需要设置注释(表'%s').",
 	ER_TABLE_MUST_HAVE_PK:                  "表 '%s' 需要设置主键.",
@@ -478,13 +489,17 @@ var ErrorsChinese = map[int]string{
 	ER_WRONG_TABLE_NAME:                    "不正确的表名: '%-.100s'",
 	ER_CANT_SET_CHARSET:                    "禁止指定字符集: '%s'",
 	ER_CANT_SET_COLLATION:                  "禁止指定排序规则: '%s'",
-	ER_MUST_HAVE_COLUMNS:                   "表至少需要有一个列.",
+	ER_MUST_AT_LEAST_ONE_COLUMN:            "表至少需要有一个列.",
+	ER_MUST_HAVE_COLUMNS:                   "表必须包含以下列: '%s'.",
 	ER_PRIMARY_CANT_HAVE_NULL:              "主键的所有列必须为NOT NULL,如需要NULL列,请改用唯一索引",
 	ErrCantRemoveAllFields:                 "禁止删除表的所有列.",
 	ErrNotFoundTableInfo:                   "没有表结构信息,跳过备份.",
 	ErrNotFoundThreadId:                    "MariaDB v%d 对回滚支持不完美,请注意确认回滚语句是否正确",
 	ErrNotFoundMasterStatus:                "无法获取master binlog信息.",
 	ErrNonUniqTable:                        "表名或别名: '%-.192s' 不唯一.",
+	ErrDataTooLong:                         "数据过长!(列 '%s',行 '%d')",
+	ErrCharsetNotSupport:                   "允许的字符集: '%s'.",
+	ErrCollationNotSupport:                 "允许的排序规则: '%s'.",
 	ErrWrongUsage:                          "%s子句无法使用%s",
 }
 
@@ -500,6 +515,7 @@ func GetErrorLevel(errorNo int) uint8 {
 		ER_NOT_ALLOWED_NULLABLE,
 		ER_TOO_MANY_KEY_PARTS,
 		ER_UDPATE_TOO_MUCH_ROWS,
+		ER_INSERT_TOO_MUCH_ROWS,
 		ER_TOO_MANY_KEYS,
 		ER_PK_TOO_MANY_PARTS,
 		ER_PK_COLS_NOT_INT,
@@ -512,6 +528,7 @@ func GetErrorLevel(errorNo int) uint8 {
 		ER_IDENT_USE_KEYWORD,
 		ER_TABLE_CHARSET_MUST_UTF8,
 		ER_TABLE_CHARSET_MUST_NULL,
+		ErrTableCollationNotSupport,
 		ER_AUTO_INCR_ID_WARNING,
 		ER_ALTER_TABLE_ONCE,
 		ER_BLOB_CANT_HAVE_DEFAULT,
@@ -533,6 +550,7 @@ func GetErrorLevel(errorNo int) uint8 {
 		ER_CANT_SET_COLLATION,
 		ErrNotFoundTableInfo,
 		ErrNotFoundThreadId,
+		ER_MUST_HAVE_COLUMNS,
 		ER_CHANGE_COLUMN_TYPE:
 		return 1
 
@@ -587,7 +605,8 @@ func GetErrorLevel(errorNo int) uint8 {
 		ER_CANT_DROP_FIELD_OR_KEY,
 		ER_NOT_SUPPORTED_YET,
 		ER_TABLE_MUST_INNODB,
-		ER_NAMES_MUST_UTF8,
+		ErrCharsetNotSupport,
+		ErrCollationNotSupport,
 		ER_FOREIGN_KEY,
 		ER_INCEPTION_EMPTY_QUERY:
 		return 2