+++ session/session.go
@@ -204,8 +204,15 @@ type session struct {
 	// 数据库版本号
 	DBVersion int
 
+	// 远程数据库线程ID,在启用备份功能时,用以记录线程ID来解析binlog
+	threadID uint32
+
+	sqlFingerprint map[string]*Record
 	// // osc进程解析通道
 	// chanOsc chan *ChanOscData
+
+	// 当前阶段 [Check,Execute,Backup]
+	stage byte
 }
 
 // DDLOwnerChecker returns s.ddlOwnerChecker.