+++ session/session_inception_test.go
@@ -43,6 +43,11 @@ type testSessionIncSuite struct {
 	store     kv.Storage
 	dom       *domain.Domain
 	tk        *testkit.TestKit
+
+	version int
+	sqlMode string
+
+	rows [][]interface{}
 }
 
 func (s *testSessionIncSuite) SetUpSuite(c *C) {
@@ -69,6 +74,8 @@ func (s *testSessionIncSuite) SetUpSuite(c *C) {
 	// config.GetGlobalConfig().Inc.Lang = "zh-CN"
 	// session.SetLanguage("zh-CN")
 	config.GetGlobalConfig().Inc.Lang = "en-US"
+	config.GetGlobalConfig().Inc.EnableFingerprint = true
+	config.GetGlobalConfig().Inc.SqlSafeUpdates = 0
 	session.SetLanguage("en-US")
 }
 
@@ -110,7 +117,7 @@ use test_inc;
 %s;
 inception_magic_commit;`
 	for _, name := range strings.Split(sql.(string), "\n") {
-		if strings.HasPrefix(name, "show tables:") {
+		if strings.HasPrefix(name, "show tables") {
 			continue
 		}
 		n := strings.Replace(name, "'", "", -1)
@@ -130,6 +137,10 @@ inception_magic_commit;`
 func (s *testSessionIncSuite) getDBVersion(c *C) int {
 	if testing.Short() {
 		c.Skip("skipping test; in TRAVIS mode")
+	}
+
+	if s.version > 0 {
+		return s.version
 	}
 
 	if s.tk == nil {
@@ -160,10 +171,40 @@ func (s *testSessionIncSuite) getDBVersion(c *C) int {
 		if err != nil {
 			fmt.Println(err)
 		}
+		s.version = version
 		return version
 	}
 
 	return 50700
+}
+
+func (s *testSessionIncSuite) getSQLMode(c *C) string {
+	if testing.Short() {
+		c.Skip("skipping test; in TRAVIS mode")
+	}
+
+	if s.sqlMode != "" {
+		return s.sqlMode
+	}
+
+	if s.tk == nil {
+		s.tk = testkit.NewTestKitWithInit(c, s.store)
+	}
+
+	sql := "show variables like 'sql_mode'"
+
+	res := makeSQL(s.tk, sql)
+	c.Assert(int(s.tk.Se.AffectedRows()), Equals, 2)
+
+	row := res.Rows()[int(s.tk.Se.AffectedRows())-1]
+	versionStr := row[5].(string)
+
+	versionStr = strings.SplitN(versionStr, "|", 2)[1]
+	value := strings.Replace(versionStr, "'", "", -1)
+	value = strings.TrimSpace(value)
+
+	s.sqlMode = value
+	return value
 }
 
 func makeSQL(tk *testkit.TestKit, sql string) *testkit.Result {
@@ -173,6 +214,22 @@ use test_inc;
 %s;
 inception_magic_commit;`
 	return tk.MustQueryInc(fmt.Sprintf(a, sql))
+}
+
+func (s *testSessionIncSuite) execSQL(c *C, sql string) {
+
+	config.GetGlobalConfig().Inc.EnableDropTable = true
+
+	a := `/*--user=test;--password=test;--host=127.0.0.1;--execute=1;--backup=0;--port=3306;--enable-ignore-warnings;*/
+inception_magic_start;
+use test_inc;
+%s;
+inception_magic_commit;`
+	res := s.tk.MustQueryInc(fmt.Sprintf(a, sql))
+
+	for _, row := range res.Rows() {
+		c.Assert(row[2], Not(Equals), "2", Commentf("%v", row))
+	}
 }
 
 func (s *testSessionIncSuite) testErrorCode(c *C, sql string, errors ...*session.SQLError) {
@@ -198,10 +255,23 @@ func (s *testSessionIncSuite) testErrorCode(c *C, sql string, errors ...*session
 		for _, e := range errors {
 			errMsgs = append(errMsgs, e.Error())
 		}
-		c.Assert(row[4], Equals, strings.Join(errMsgs, "\n"))
+		c.Assert(row[4], Equals, strings.Join(errMsgs, "\n"), Commentf("%v", row))
 	}
 
-	c.Assert(row[2], Equals, strconv.Itoa(errCode))
+	c.Assert(row[2], Equals, strconv.Itoa(errCode), Commentf("%v", row))
+
+	s.rows = res.Rows()
+}
+
+func (s *testSessionIncSuite) testAffectedRows(c *C, affectedRows ...int) {
+	if len(s.rows) == 0 {
+		return
+	}
+	count := len(affectedRows)
+	for i, affectedRow := range affectedRows {
+		row := s.rows[len(s.rows)-(count-i)]
+		c.Assert(row[6], Equals, strconv.Itoa(affectedRow), Commentf("%v", row))
+	}
 }
 
 func (s *testSessionIncSuite) TestBegin(c *C) {
@@ -306,22 +376,6 @@ func (s *testSessionIncSuite) TestCreateTable(c *C) {
 	sql = `create table t1(id int,c1 char(200));`
 	s.testErrorCode(c, sql,
 		session.NewErr(session.ER_CHAR_TO_VARCHAR_LEN, "c1"))
-
-	// 字符集
-	sql = `create table t1(id int,c1 varchar(20) character set utf8,
-		c2 varchar(20) COLLATE utf8_bin);`
-	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_CHARSET_ON_COLUMN, "t1", "c1"),
-		session.NewErr(session.ER_CHARSET_ON_COLUMN, "t1", "c2"))
-
-	config.GetGlobalConfig().Inc.EnableSetCharset = false
-	sql = `create table t1(id int,c1 varchar(20)) character set utf8;`
-	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_TABLE_CHARSET_MUST_NULL, "t1"))
-
-	sql = `create table t1(id int,c1 varchar(20)) COLLATE utf8_bin;`
-	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_TABLE_CHARSET_MUST_NULL, "t1"))
 
 	// 关键字
 	config.GetGlobalConfig().Inc.EnableIdentiferKeyword = false
@@ -470,7 +524,7 @@ func (s *testSessionIncSuite) TestCreateTable(c *C) {
 	config.GetGlobalConfig().Inc.SupportCharset = "utf8mb4"
 	sql = "create table t1(a int) character set utf8;"
 	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_NAMES_MUST_UTF8, "utf8mb4"))
+		session.NewErr(session.ErrCharsetNotSupport, "utf8mb4"))
 
 	config.GetGlobalConfig().Inc.EnableSetCharset = true
 	config.GetGlobalConfig().Inc.SupportCharset = "utf8,utf8mb4"
@@ -481,7 +535,7 @@ func (s *testSessionIncSuite) TestCreateTable(c *C) {
 	config.GetGlobalConfig().Inc.SupportCharset = "utf8,utf8mb4"
 	sql = "create table t1(a int) character set laitn1;"
 	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_NAMES_MUST_UTF8, "utf8,utf8mb4"))
+		session.NewErr(session.ErrCharsetNotSupport, "utf8,utf8mb4"))
 
 	// 外键
 	sql = "create table test_error_code (a int not null ,b int not null,c int not null, d int not null, foreign key (b, c) references product(id));"
@@ -495,11 +549,11 @@ func (s *testSessionIncSuite) TestCreateTable(c *C) {
 
 	sql = "create table test_error_code_2;"
 	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_MUST_HAVE_COLUMNS))
+		session.NewErr(session.ER_MUST_AT_LEAST_ONE_COLUMN))
 
 	sql = "create table test_error_code_2 (unique(c1));"
 	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_MUST_HAVE_COLUMNS))
+		session.NewErr(session.ER_MUST_AT_LEAST_ONE_COLUMN))
 
 	sql = "create table test_error_code_2(c1 int, c2 int, c3 int, primary key(c1), primary key(c2));"
 	s.testErrorCode(c, sql,
@@ -635,6 +689,28 @@ primary key(id)) comment 'test';`
 
 	config.GetGlobalConfig().Inc.EnableNullable = true
 	sql = `drop table if exists t1;CREATE TABLE t1(c1 int);`
+	s.testErrorCode(c, sql)
+
+	// 检查必须的字段
+	config.GetGlobalConfig().Inc.MustHaveColumns = "c1"
+	sql = `drop table if exists t1;CREATE TABLE t1(id int);`
+	s.testErrorCode(c, sql,
+		session.NewErr(session.ER_MUST_HAVE_COLUMNS, "c1"))
+
+	sql = `drop table if exists t1;CREATE TABLE t1(c1 int);`
+	s.testErrorCode(c, sql)
+
+	config.GetGlobalConfig().Inc.MustHaveColumns = "c1 int,c2 datetime"
+
+	sql = `drop table if exists t1;CREATE TABLE t1(id int);`
+	s.testErrorCode(c, sql,
+		session.NewErr(session.ER_MUST_HAVE_COLUMNS, "c1 int,c2 datetime"))
+
+	sql = `drop table if exists t1;CREATE TABLE t1(c1 bigint,c2 int);`
+	s.testErrorCode(c, sql,
+		session.NewErr(session.ER_MUST_HAVE_COLUMNS, "c1 int,c2 datetime"))
+
+	sql = `drop table if exists t1;CREATE TABLE t1(c1 int,c2 datetime);`
 	s.testErrorCode(c, sql)
 
 }
@@ -1086,6 +1162,33 @@ func (s *testSessionIncSuite) TestInsert(c *C) {
 	sql = "create table t1(c1 char(100) not null);insert into t1(c1) select t1.c1 from t1 order by 1 union all select t1.c1 from t1;"
 	s.testErrorCode(c, sql,
 		session.NewErr(session.ErrWrongUsage, "UNION", "ORDER BY"))
+
+	// insert 行数
+	config.GetGlobalConfig().Inc.MaxInsertRows = 1
+
+	sql = `drop table if exists t1;create table t1(id int);
+insert into t1 values(1);`
+	s.testErrorCode(c, sql)
+	s.testAffectedRows(c, 1)
+
+	sql = `drop table if exists t1;create table t1(id int);
+insert into t1 values(1),(2);`
+	s.testErrorCode(c, sql,
+		session.NewErr(session.ER_INSERT_TOO_MUCH_ROWS, 2, 1))
+
+	config.GetGlobalConfig().Inc.MaxInsertRows = 3
+	sql = `drop table if exists t1;create table t1(id int);
+insert into t1 values(1),(2),(3);`
+
+	s.testErrorCode(c, sql)
+	s.testAffectedRows(c, 3)
+
+	s.execSQL(c, "drop table if exists t1;create table t1(id int);")
+	sql = `drop table if exists t2;create table t2 like t1;
+insert into t2 select id from t1;`
+	s.testErrorCode(c, sql)
+	s.testAffectedRows(c, 1)
+
 }
 
 func (s *testSessionIncSuite) TestUpdate(c *C) {
@@ -1173,6 +1276,22 @@ func (s *testSessionIncSuite) TestUpdate(c *C) {
 		update table1 t1,table2 t2 set t1.c1=t2.c2 where t1.id1=t2.id2;`
 	s.testErrorCode(c, sql)
 
+	sql = `drop table if exists table1;drop table if exists table2;
+		create table table1(id1 int primary key,c1 int);
+		create table table2(id2 int primary key,c2 int,c22 int);
+		update table1 a1,table2 a2 set a1.c1=a2.c2 where a1.id1=a2.id2 and a1.c1=a2.c2 and a1.id1 in (1,2,3);`
+	s.testErrorCode(c, sql)
+
+	sql = `drop table if exists table1;drop table if exists table2;
+		create table table1(id1 int primary key,c1 int);
+		create table test.table2(id2 int primary key,c2 int,c22 int);
+		update table1 a1,test.table2 a2 set a1.c1=a2.c2 where a1.id1=a2.id2 and a1.c1=a2.c2 and a1.id1 in (1,2,3);`
+	s.testErrorCode(c, sql)
+
+	s.execSQL(c, "drop table if exists t1;create table t1(id int,c1 int);insert into t1(id) values(1);")
+	sql = `update t1 set c1=1 where id =1;`
+	s.testErrorCode(c, sql)
+	s.testAffectedRows(c, 1)
 }
 
 func (s *testSessionIncSuite) TestDelete(c *C) {
@@ -1245,6 +1364,11 @@ func (s *testSessionIncSuite) TestDelete(c *C) {
 	row := res.Rows()[int(tk.Se.AffectedRows())-1]
 	c.Assert(row[2], Equals, "0")
 	c.Assert(row[6], Equals, "0")
+
+	s.execSQL(c, "drop table if exists t1;create table t1(id int,c1 int);insert into t1(id) values(1);")
+	sql = `delete from t1 where id =1;`
+	s.testErrorCode(c, sql)
+	s.testAffectedRows(c, 1)
 }
 
 func (s *testSessionIncSuite) TestCreateDataBase(c *C) {
@@ -1295,8 +1419,7 @@ func (s *testSessionIncSuite) TestCreateDataBase(c *C) {
 	config.GetGlobalConfig().Inc.SupportCharset = "utf8mb4"
 	sql = "drop database test1;create database test1 character set utf8;"
 	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_CANT_SET_CHARSET, "utf8"),
-		session.NewErr(session.ER_NAMES_MUST_UTF8, "utf8mb4"))
+		session.NewErr(session.ER_CANT_SET_CHARSET, "utf8"))
 
 	config.GetGlobalConfig().Inc.EnableSetCharset = true
 	config.GetGlobalConfig().Inc.SupportCharset = "utf8,utf8mb4"
@@ -1307,11 +1430,13 @@ func (s *testSessionIncSuite) TestCreateDataBase(c *C) {
 	config.GetGlobalConfig().Inc.SupportCharset = "utf8,utf8mb4"
 	sql = "drop database test1;create database test1 character set laitn1;"
 	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_NAMES_MUST_UTF8, "utf8,utf8mb4"))
+		session.NewErr(session.ErrCharsetNotSupport, "utf8,utf8mb4"))
 }
 
 func (s *testSessionIncSuite) TestTimestampColumn(c *C) {
 	sql := ""
+
+	fmt.Println("SQLMode: ", s.getSQLMode(c))
 
 	sql = `drop table if exists timeTable;create table timeTable(c1 timestamp default '');`
 	s.testErrorCode(c, sql,
@@ -1319,19 +1444,11 @@ func (s *testSessionIncSuite) TestTimestampColumn(c *C) {
 	sql = `drop table if exists timeTable;create table timeTable(c1 timestamp default '0');`
 	s.testErrorCode(c, sql,
 		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
-	sql = `drop table if exists timeTable;create table timeTable(c1 timestamp default '0000-0-0');`
-	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
-	sql = `drop table if exists timeTable;create table timeTable(c1 timestamp default '0000-0-0 00:00:00');`
-	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
+
 	sql = `drop table if exists timeTable;create table timeTable(c1 timestamp default '2100-1-1 1:1:1');`
 	s.testErrorCode(c, sql,
 		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
 	sql = `drop table if exists timeTable;create table timeTable(c1 timestamp default '1900-1-1 1:1:1');`
-	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
-	sql = `drop table if exists timeTable;create table timeTable(c1 timestamp default '2000-1-0 1:1:1');`
 	s.testErrorCode(c, sql,
 		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
 
@@ -1339,15 +1456,6 @@ func (s *testSessionIncSuite) TestTimestampColumn(c *C) {
 	s.testErrorCode(c, sql,
 		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
 	sql = `drop table if exists timeTable;create table timeTable(c1 datetime default '0');`
-	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
-	sql = `drop table if exists timeTable;create table timeTable(c1 datetime default '0000-0-0');`
-	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
-	sql = `drop table if exists timeTable;create table timeTable(c1 datetime default '0000-0-0 00:00:00');`
-	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
-	sql = `drop table if exists timeTable;create table timeTable(c1 datetime default '2000-1-0 1:1:1');`
 	s.testErrorCode(c, sql,
 		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
 
@@ -1360,15 +1468,6 @@ func (s *testSessionIncSuite) TestTimestampColumn(c *C) {
 	s.testErrorCode(c, sql,
 		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
 	sql = `drop table if exists timeTable;create table timeTable(c1 date default '0');`
-	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
-	sql = `drop table if exists timeTable;create table timeTable(c1 date default '0000-0-0');`
-	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
-	sql = `drop table if exists timeTable;create table timeTable(c1 date default '0000-0-0 00:00:00');`
-	s.testErrorCode(c, sql,
-		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
-	sql = `drop table if exists timeTable;create table timeTable(c1 date default '2000-1-0 1:1:1');`
 	s.testErrorCode(c, sql,
 		session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
 
@@ -1391,6 +1490,71 @@ func (s *testSessionIncSuite) TestTimestampColumn(c *C) {
 	sql = `drop table if exists timeTable;create table timeTable(c1 timestamp(6) default '2000-1-1 1:1:1');`
 	s.testErrorCode(c, sql)
 
+	// 零值审核
+	if strings.Contains(s.getSQLMode(c), "TRADITIONAL") ||
+		(strings.Contains(s.getSQLMode(c), "STRICT_") && strings.Contains(s.getSQLMode(c), "NO_ZERO_DATE")) {
+		sql = `drop table if exists timeTable;create table timeTable(c1 timestamp default '0000-0-0');`
+		s.testErrorCode(c, sql,
+			session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
+		sql = `drop table if exists timeTable;create table timeTable(c1 timestamp default '0000-0-0 00:00:00');`
+		s.testErrorCode(c, sql,
+			session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
+
+		sql = `drop table if exists timeTable;create table timeTable(c1 datetime default '0000-0-0');`
+		s.testErrorCode(c, sql,
+			session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
+		sql = `drop table if exists timeTable;create table timeTable(c1 datetime default '0000-0-0 00:00:00');`
+		s.testErrorCode(c, sql,
+			session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
+
+		sql = `drop table if exists timeTable;create table timeTable(c1 date default '0000-0-0');`
+		s.testErrorCode(c, sql,
+			session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
+		sql = `drop table if exists timeTable;create table timeTable(c1 date default '0000-0-0 00:00:00');`
+		s.testErrorCode(c, sql,
+			session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
+	} else {
+		sql = `drop table if exists timeTable;create table timeTable(c1 timestamp default '0000-0-0');`
+		s.testErrorCode(c, sql)
+		sql = `drop table if exists timeTable;create table timeTable(c1 timestamp default '0000-0-0 00:00:00');`
+		s.testErrorCode(c, sql)
+
+		sql = `drop table if exists timeTable;create table timeTable(c1 datetime default '0000-0-0');`
+		s.testErrorCode(c, sql)
+		sql = `drop table if exists timeTable;create table timeTable(c1 datetime default '0000-0-0 00:00:00');`
+		s.testErrorCode(c, sql)
+
+		sql = `drop table if exists timeTable;create table timeTable(c1 date default '0000-0-0');`
+		s.testErrorCode(c, sql)
+		sql = `drop table if exists timeTable;create table timeTable(c1 date default '0000-0-0 00:00:00');`
+		s.testErrorCode(c, sql)
+	}
+
+	// 月,日零值审核
+	if strings.Contains(s.getSQLMode(c), "TRADITIONAL") ||
+		(strings.Contains(s.getSQLMode(c), "STRICT_") && strings.Contains(s.getSQLMode(c), "NO_ZERO_IN_DATE")) {
+		sql = `drop table if exists timeTable;create table timeTable(c1 timestamp default '2000-1-0 1:1:1');`
+		s.testErrorCode(c, sql,
+			session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
+
+		sql = `drop table if exists timeTable;create table timeTable(c1 datetime default '2000-1-0 1:1:1');`
+		s.testErrorCode(c, sql,
+			session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
+
+		sql = `drop table if exists timeTable;create table timeTable(c1 date default '2000-1-0 1:1:1');`
+		s.testErrorCode(c, sql,
+			session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
+	} else {
+		sql = `drop table if exists timeTable;create table timeTable(c1 timestamp default '2000-1-0 1:1:1');`
+		s.testErrorCode(c, sql,
+			session.NewErr(session.ER_INVALID_DEFAULT, "c1"))
+
+		sql = `drop table if exists timeTable;create table timeTable(c1 datetime default '2000-1-0 1:1:1');`
+		s.testErrorCode(c, sql)
+
+		sql = `drop table if exists timeTable;create table timeTable(c1 date default '2000-1-0 1:1:1');`
+		s.testErrorCode(c, sql)
+	}
 }
 
 func (s *testSessionIncSuite) TestRenameTable(c *C) {
@@ -1454,4 +1618,140 @@ func (s *testSessionIncSuite) TestAlterTableDropIndex(c *C) {
 
 	sql = "create table t1(c1 int);alter table t1 add index idx (c1);alter table t1 drop index idx;"
 	s.testErrorCode(c, sql)
+}
+
+func (s *testSessionIncSuite) TestAlterTable(c *C) {
+	saved := config.GetGlobalConfig().Inc
+	defer func() {
+		config.GetGlobalConfig().Inc = saved
+	}()
+
+	config.GetGlobalConfig().Inc.CheckColumnComment = false
+	config.GetGlobalConfig().Inc.CheckTableComment = false
+	config.GetGlobalConfig().Inc.EnableDropTable = true
+
+	// 删除后添加列
+	sql = "drop table if exists t1;create table t1(id int,c1 int);alter table t1 drop column c1;alter table t1 add column c1 varchar(20);"
+	s.testErrorCode(c, sql)
+
+	sql = "drop table if exists t1;create table t1(id int,c1 int);alter table t1 drop column c1,add column c1 varchar(20);"
+	s.testErrorCode(c, sql)
+
+	// 删除后添加索引
+	sql = "drop table if exists t1;create table t1(id int,c1 int,key ix(c1));alter table t1 drop index ix;alter table t1 add index ix(c1);"
+	s.testErrorCode(c, sql)
+
+	sql = "drop table if exists t1;create table t1(id int,c1 int,key ix(c1));alter table t1 drop index ix,add index ix(c1);"
+	s.testErrorCode(c, sql)
+
+}
+
+func (s *testSessionIncSuite) TestCreateTablePrimaryKey(c *C) {
+	saved := config.GetGlobalConfig().Inc
+	defer func() {
+		config.GetGlobalConfig().Inc = saved
+	}()
+
+	sql := ""
+
+	config.GetGlobalConfig().Inc.CheckColumnComment = false
+	config.GetGlobalConfig().Inc.CheckTableComment = false
+
+	// EnablePKColumnsOnlyInt
+	config.GetGlobalConfig().Inc.EnablePKColumnsOnlyInt = true
+
+	sql = "create table t1(id tinyint, primary key(id));"
+	s.testErrorCode(c, sql,
+		session.NewErr(session.ER_PK_COLS_NOT_INT, "id", "test_inc", "t1"))
+
+	sql = "create table t1(id mediumint, primary key(id));"
+	s.testErrorCode(c, sql)
+
+	sql = "create table t1(id int, primary key(id));"
+	s.testErrorCode(c, sql)
+
+	sql = "create table t1(id bigint, primary key(id));"
+	s.testErrorCode(c, sql)
+
+	sql = "create table t1(id varchar(10), primary key(id));"
+	s.testErrorCode(c, sql,
+		session.NewErr(session.ER_PK_COLS_NOT_INT, "id", "test_inc", "t1"))
+
+	sql = "create table t1(id tinyint primary key);"
+	s.testErrorCode(c, sql,
+		session.NewErr(session.ER_PK_COLS_NOT_INT, "id", "test_inc", "t1"))
+
+	sql = "create table t1(id mediumint primary key);"
+	s.testErrorCode(c, sql)
+
+	sql = "create table t1(id int primary key);"
+	s.testErrorCode(c, sql)
+
+	sql = "create table t1(id bigint primary key);"
+	s.testErrorCode(c, sql)
+
+	sql = "create table t1(id varchar(10) primary key);"
+	s.testErrorCode(c, sql,
+		session.NewErr(session.ER_PK_COLS_NOT_INT, "id", "test_inc", "t1"))
+}
+
+func (s *testSessionIncSuite) TestTableCharsetCollation(c *C) {
+	tk := testkit.NewTestKitWithInit(c, s.store)
+	saved := config.GetGlobalConfig().Inc
+	defer func() {
+		config.GetGlobalConfig().Inc = saved
+	}()
+
+	sql := ""
+
+	config.GetGlobalConfig().Inc.CheckColumnComment = false
+	config.GetGlobalConfig().Inc.CheckTableComment = false
+
+	// 表存在
+	res := makeSQL(tk, "create table t1(id int);create table t1(id int);")
+	row := res.Rows()[int(tk.Se.AffectedRows())-1]
+	c.Assert(row[2], Equals, "2")
+	c.Assert(row[4], Equals, "Table 't1' already exists.")
+
+	// 字符集
+	sql = `create table t1(id int,c1 varchar(20) character set utf8,
+		c2 varchar(20) COLLATE utf8_bin);`
+	s.testErrorCode(c, sql,
+		session.NewErr(session.ER_CHARSET_ON_COLUMN, "t1", "c1"),
+		session.NewErr(session.ER_CHARSET_ON_COLUMN, "t1", "c2"))
+
+	config.GetGlobalConfig().Inc.EnableSetCharset = false
+	sql = `create table t1(id int,c1 varchar(20)) character set utf8;`
+	s.testErrorCode(c, sql,
+		session.NewErr(session.ER_TABLE_CHARSET_MUST_NULL, "t1"))
+
+	config.GetGlobalConfig().Inc.EnableSetCollation = false
+	sql = `create table t1(id int,c1 varchar(20)) COLLATE utf8_bin;`
+	s.testErrorCode(c, sql,
+		session.NewErr(session.ErrTableCollationNotSupport, "t1"))
+
+	sql = `create table t1(id int,c1 varchar(20)) character set utf8 COLLATE utf8_bin;`
+	s.testErrorCode(c, sql,
+		session.NewErr(session.ER_TABLE_CHARSET_MUST_NULL, "t1"),
+		session.NewErr(session.ErrTableCollationNotSupport, "t1"))
+
+	config.GetGlobalConfig().Inc.EnableSetCharset = true
+	config.GetGlobalConfig().Inc.EnableSetCollation = true
+	config.GetGlobalConfig().Inc.SupportCharset = ""
+	config.GetGlobalConfig().Inc.SupportCollation = ""
+
+	sql = `create table t1(id int,c1 varchar(20)) character set utf8 COLLATE utf8_bin;`
+	s.testErrorCode(c, sql)
+
+	config.GetGlobalConfig().Inc.SupportCharset = "utf8"
+	sql = `create table t1(id int,c1 varchar(20)) character set utf8mb4 COLLATE utf8_bin;`
+	s.testErrorCode(c, sql,
+		session.NewErr(session.ErrCharsetNotSupport, "utf8"))
+
+	config.GetGlobalConfig().Inc.SupportCollation = "utf8_bin"
+	sql = `create table t1(id int,c1 varchar(20)) character set utf8mb4 COLLATE utf8mb4_bin;`
+	s.testErrorCode(c, sql,
+		session.NewErr(session.ErrCharsetNotSupport, "utf8"),
+		session.NewErr(session.ErrCollationNotSupport, "utf8_bin"))
+
 }